+++ distribution/release-notes.md
@@ -4,13 +4,14 @@
 
 - Pirate missions will now calculate correctly based on bounty collected against a target faction. Old bounties no longer count towards the mission. (bug fix)
 - Pirate mission accept would fail if no hunting were available. fixed.
-- NOTE: If you cash all your bounties before completing the missions, the counter will reset. (known issue)
 - Added Windows key support for bindings.
 - Added application menu support for key bindings.
 - Found and removed long and unnecessary reminder in pirate missions.
 - Say "Exit" or "Close" to exit any window or tab.
 - Nightvision toggle no longer requires on/off clarification.
 - Lights toggle no longer requires on/off clarification.
+- Reminders can now be interrupted
+- Added better support for subsystem targeting. Say "target fsd" or "target drive" or "target engines" or "target power distributor" etc. NOTE: Optional subsystems are not included.
 
 ## Elite Intel v-0.0333-beta